+++ service/src/main/java/org/grnet/cat/services/assessment/AssessmentService.java
@@ -3,21 +3,22 @@ package org.grnet.cat.services.assessment;
 import jakarta.ws.rs.core.UriInfo;
 import org.grnet.cat.dtos.assessment.AssessmentResponse;
 import org.grnet.cat.dtos.pagination.PageResource;
-import org.grnet.cat.entities.Assessment;
 
-public interface AssessmentService <Request extends AssessmentRequest, Update, Response extends AssessmentResponse, Entity> {
+public interface AssessmentService <Request, Update, Response extends AssessmentResponse, Entity> {
 
     Response createAssessment(String userId, Request request);
 
     Response getDtoAssessment(String userId, String id);
 
+    Response getPublicDtoAssessment(String id);
+
     void deleteAll();
 
     Response update(String id, String userId, Update request);
 
-    PageResource<? extends Response> getDtoAssessmentsByUserAndPage(int page, int size, UriInfo uriInfo, String userID);
+    PageResource<? extends Response> getDtoAssessmentsByUserAndPage(int page, int size, UriInfo uriInfo, String userID, String subjectName, String subjectType, Long actorId);
 
-    PageResource<? extends Response> getPublishedDtoAssessmentsByTypeAndActorAndPage(int page, int size, Long typeId, Long actorId, UriInfo uriInfo);
+    PageResource<? extends Response> getPublishedDtoAssessmentsByTypeAndActorAndPage(int page, int size, Long typeId, Long actorId, UriInfo uriInfo, String subjectName, String subjectType);
 
     void assessmentBelongsToUser(String userID, String id);